# agentflow-server/src/main/resources/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AgentFlow API",
    "version": "0.1.0",
    "description": "与当前工程同步的已实现 API。演示认证尚非企业身份服务；Agent、财务支付、外部 Webhook、组织和附件接口尚未交付。所有业务资源限定认证主体租户，客户端不能指定他人租户绕过授权。",
    "contact": {
      "email": "owlzhangfq@gmail.com"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "当前部署同源根路径；paths 已包含 /api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/process-definitions": {
      "get": {
        "operationId": "listDefinitions",
        "tags": [
          "流程设计"
        ],
        "summary": "流程版本列表",
        "description": "管理角色可读本租户全部定义，其他角色只读已发布定义。",
        "x-access": "当前已登录账号",
        "x-idempotency": false,
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Definition"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "角色或当前资源权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "领域规则、表单、流程图或状态校验不通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "未列出的 HTTP 或基础设施错误；客户端不能把结果未知的写入当作失败后创建新请求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "状态；非管理角色指定非 PUBLISHED 返回 403",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "ARCHIVED"
              ]
            }
          }
        ]
      },
      "post": {
        "operationId": "createDefinition",
        "tags": [
          "流程设计"
        ],
        "summary": "创建流程草稿",
        "description": "创建独立草稿，version=0；发布后不可原地修改。",
        "x-access": "ADMIN 或 PROCESS_ADMIN",
        "x-idempotency": true,
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Definition"
                }
              }
            },
            "headers": {
              "Idempotency-Replayed": {
                "description": "true 为原成功响应回放；false 为本次执行。",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "请求格式、筛选参数或幂等键无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "角色或当前资源权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本冲突、定义绑定冲突、委派状态冲突或幂等键复用/过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "领域规则、表单、流程图或状态校验不通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "未列出的 HTTP 或基础设施错误；客户端不能把结果未知的写入当作失败后创建新请求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "每个新写入唯一。同一租户内绑定账号、角色、方法、路径、原始查询与正文；网络结果不明时使用原请求逐字节重试。成功响应保存 24 小时，过期拒绝重放，不重新执行。",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDefinition"
              },
              "example": {
                "key": "integration-example",
                "name": "集成示例流程",
                "graph": {
                  "nodes": [
                    {
                      "id": "start",
                      "name": "开始",
                      "type": "START",
                      "properties": {}
                    },
                    {
                      "id": "review",
                      "name": "部门审批",
                      "type": "USER_TASK",
                      "properties": {
                        "assigneeRule": "role:MANAGER"
                      }
                    },
                    {
                      "id": "end",
                      "name": "结束",
                      "type": "END",
                      "properties": {}
                    }
                  ],
                  "edges": [
                    {
                      "id": "first",
                      "source": "start",
                      "target": "review",
                      "condition": "",
                      "defaultBranch": false
                    },
                    {
                      "id": "last",
                      "source": "review",
                      "target": "end",
                      "condition": "",
                      "defaultBranch": false
                    }
                  ]
                },
                "formSchema": null
              }
            }
          }
        }
      }
    },
    "/api/v1/process-definitions/{id}": {
      "get": {
        "operationId": "getDefinition",
        "tags": [
          "流程设计"
        ],
        "summary": "读取流程定义",
        "description": "普通账号读取草稿返回 404。",
        "x-access": "当前已登录账号",
        "x-idempotency": false,
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Definition"
                }
              }
            }
          },
          "400": {
            "description": "请求格式、筛选参数或幂等键无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "未列出的 HTTP 或基础设施错误；客户端不能把结果未知的写入当作失败后创建新请求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "资源标识",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      },
      "put": {
        "operationId": "updateDefinition",
        "tags": [
          "流程设计"
        ],
        "summary": "保存草稿",
        "description": "expectedRevision 来自当前读取。冲突后重新加载、核对，不能自动用新版本覆盖。null formSchema 保留旧表单。",
        "x-access": "ADMIN 或 PROCESS_ADMIN",
        "x-idempotency": true,
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Definition"
                }
              }
            },
            "headers": {
              "Idempotency-Replayed": {
                "description": "true 为原成功响应回放；false 为本次执行。",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "请求格式、筛选参数或幂等键无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "角色或当前资源权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本冲突、定义绑定冲突、委派状态冲突或幂等键复用/过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "领域规则、表单、流程图或状态校验不通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "未列出的 HTTP 或基础设施错误；客户端不能把结果未知的写入当作失败后创建新请求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "资源标识",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "每个新写入唯一。同一租户内绑定账号、角色、方法、路径、原始查询与正文；网络结果不明时使用原请求逐字节重试。成功响应保存 24 小时，过期拒绝重放，不重新执行。",
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._:-]{1,128}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDefinition"
              },
              "example": {
                "name": "集成示例流程",
                "graph": {
                  "nodes": [
                    {
                      "id": "start",
                      "name": "开始",
                      "type": "START",
                      "properties": {}
                    },
                    {
                      "id": "review",
                      "name": "部门审批",
                      "type": "USER_TASK",
                      "properties": {
                        "assigneeRule": "role:MANAGER"
                      }
                    },
                    {
                      "id": "end",
                      "name": "结束",
                      "type": "END",
                      "properties": {}
                    }
                  ],
                  "edges": [
                    {
                      "id": "first",
                      "source": "start",
                      "target": "review",
                      "condition": "",
                      "defaultBranch": false
                    },
                    {
                      "id": "last",
                      "source": "review",
                      "target": "end",
                      "condition": "",
                      "defaultBranch": false
                    }
                  ]
                },
                "expectedRevision": 0,
                "formSchema": null
              }
            }
          }
        }
      }
    },
    "/api/v1/process-definitions/validate": {
      "post": {
        "operationId": "validateDefinition",
        "tags": [
          "流程设计"
        ],
        "summary": "校验流程设计",
        "description": "只读，不创建草稿。HTTP 200 的 errors 数组非空表示设计未通过；结构/表单解码失败可为 422。 管理员同时检查身份源中是否存在有效审批人；普通用户只检查结构，不开放账号存在性探测。",
        "x-access": "当前已登录账号",
        "x-idempotency": false,
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationResult"
                }
              }
            }
          },
          "400": {
            "description": "请求格式、筛选参数或幂等键无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "领域规则、表单、流程图或状态校验不通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "未列出的 HTTP 或基础设施错误；客户端不能把结果未知的写入当作失败后创建新请求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphInput"
              },
              "example": {
                "graph": {
                  "nodes": [
                    {
                      "id": "start",
                      "name": "开始",
                      "type": "START",
                      "properties": {}
                    },
                    {
                      "id": "review",
                      "name": "部门审批",
                      "type": "USER_TASK",
                      "properties": {
                        "assigneeRule": "role:MANAGER"
                      }
                    },
                    {
                      "id": "end",
                      "name": "结束",
                      "type": "END",
                      "properties": {}
                    }
                  ],
                  "edges": [
                    {
                      "id": "first",
                      "source": "start",
                      "target": "review",
                      "condition": "",
                      "defaultBranch": false
                    },
                    {
                      "id": "last",
                      "source": "review",
                      "target": "end",
                      "condition": "",
                      "defaultBranch": false
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/process-definitions/simulate": {
      "post": {
        "operationId": "previewDefinition",
        "tags": [
          "流程设计"
        ],
        "summary": "模拟当前设计",
        "description": "只读。返回真实领域求值的节点路径、连线与分支依据，不保存测试数据。",
        "x-access": "ADMIN 或 PROCESS_ADMIN",
        "x-idempotency": false,
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            }
          },
          "400": {
            "description": "请求格式、筛选参数或幂等键无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "角色或当前资源权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "领域规则、表单、流程图或状态校验不通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "未列出的 HTTP 或基础设施错误；客户端不能把结果未知的写入当作失败后创建新请求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationInput"
              },
              "example": {
                "graph": {
                  "nodes": [
                    {
                      "id": "start",
                      "name": "开始",
                      "type": "START",
                      "properties": {}
                    },
                    {
                      "id": "review",
                      "name": "部门审批",
                      "type": "USER_TASK",
                      "properties": {
                        "assigneeRule": "role:MANAGER"
                      }
                    },
                    {
                      "id": "end",
                      "name": "结束",
                      "type": "END",
                      "properties": {}
                    }
                  ],
                  "edges": [
                    {
                      "id": "first",
                      "source": "start",
                      "target": "review",
                      "condition": "",
                      "defaultBranch": false
                    },
                    {
                      "id": "last",
                      "source": "review",
                      "target": "end",
                      "condition": "",
                      "defaultBranch": false
                    }
                  ]
                },
                "values": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/process-definitions/{id}/simulate": {
      "post": {
        "operationId": "simulateSavedDefinition",
        "tags": [
          "流程设计"
        ],
        "summary": "模拟已保存版本",
        "description": "只读兼容入口，只返回 path；省略正文时使用空测试数据。按定义读取权限授权。",
        "x-access": "当前已登录账号",
        "x-idempotency": false,
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSimulationResult"
                }
              }
            }
          },
          "400": {
            "description": "请求格式、筛选参数或幂等键无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          },
          "401": {
            "description": "未登录或令牌无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不可见",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "领域规则、表单、流程图或状态校验不通过",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "未列出的 HTTP 或基础设施错误；客户端不能把结果未知的写入当作失败后创建新请求。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "资源标识",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedSimulationInput"
              },
              "example": {
                "values": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "不透明会话令牌。演示登录取得；生产身份源尚未接入。"
      }
    },
    "schemas": {
...
      "Definition": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "revision": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ]
          },
          "graph": {
            "$ref": "#/components/schemas/Graph"
          },
          "formSchema": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FormSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "tenantId",
          "key",
          "name",
          "version",
          "revision",
          "status",
          "graph",
          "formSchema"
        ]
      },
...
      "RequestError": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "$ref": "#/components/schemas/FrameworkError"
          }
        ],
        "description": "业务入口校验返回 Error；UUID 等 Spring 参数转换失败返回 FrameworkError。客户端先检查 HTTP 状态，再按是否有 code 分支。"
      },
...
    }
  }
}